fix: Resolve compilation errors in trust pipeline integration
- Fix HudState.Ammo to use AmmoClip property
- Add parentheses to ThreatLevel() and OpportunityLevel() extension method calls
- Change TrustedSuperstate.Signature from init-only to settable property
- Add explicit int cast for AverageTimeToAct float to int conversion

+++ src/OKET.Core/Cognition/ReactionPredictor.cs
@@ -83,7 +83,7 @@ public sealed class ReactionPredictor
             {
                 RecommendedReaction = learning.BestReaction,
                 Confidence = learning.Confidence * bestMatchScore,
-                TimeToAct = learning.AverageTimeToAct,
+                TimeToAct = (int)learning.AverageTimeToAct,
                 ExpectedOutcome = learning.AverageOutcome,
                 IsLearned = true,
                 ContextMatch = bestMatchScore
